EES-3798 - additional tidying up for PR. Refactoring special CSV-validating code to be reusable in Admin during Data and Metafile upload
